requirements.txt
boto3==1.28.63
botocore==1.31.63
greenlet==3.0.0
jmespath==1.0.1
numpy==1.26.0
pandas==2.1.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
